+++ philosophers.h
@@ -6,7 +6,7 @@
 /*   By: zalaksya <zalaksya@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2025/03/17 01:47:20 by zalaksya          #+#    #+#             */
-/*   Updated: 2025/06/01 12:10:56 by zalaksya         ###   ########.fr       */
+/*   Updated: 2025/06/01 19:03:00 by zalaksya         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -27,7 +27,7 @@ typedef struct s_philo
 	int				id;
 	pthread_t		thread;
 	int				meals_eaten;
-	int64_t			last_meal_time;
+	size_t		last_meal_time;
 	pthread_mutex_t	*l_fork;
 	pthread_mutex_t	*r_fork;
 	struct s_data	*data;
@@ -36,12 +36,12 @@ typedef struct s_philo
 typedef struct s_data
 {
 	int				n_philo;
-	int64_t			t_die;
-	int64_t			start_simulation;
+	size_t			t_die;
+	size_t			start_simulation;
 	int				simulation_ended;
-	int64_t			t_eat;
+	size_t			t_eat;
 	int				t_t_eat;
-	int64_t			t_sleep;
+	size_t			t_sleep;
 	int				max_meals;
 	int				someone_died;
 	pthread_mutex_t	print_lock;
